+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - default clone directory to $HOME/ghorg
 - users/orgs directory no longer appends "\_ghorg" or forces underscores
 - make $HOME/.config/ghorg/conf.yaml optional
+- color is off by default
 ### Deprecated
 ### Removed
 ### Fixed